fix: adjust messaging messages status + add testing feature (#44)
* feat: add the manual registration of messages gated under messaging_test feature

* fix: remove unused files

* fix: ensure the first nonce is 0 and use other value for sealed messages

* fix: directly use the appchain message type instead of nonce magic value

* tests: add messaging-test test

* fix: run scarb fmt

* ci: call snforge using scarb interface

* tests: fix nonce since it starts to 0 now

* fix: run scarb fmt

* fix(bindings): regenerate bindings with latest version
