+++ .github/workflows/check.yml
@@ -24,4 +24,4 @@ jobs:
     steps:
       - uses: actions/checkout@v4
       - uses: asdf-vm/actions/install@v3
-      - run: snforge test
+      - run: scarb test --all-features
