+++ .tool-versions
@@ -1,2 +1,2 @@
-scarb 2.8.0
+scarb 2.8.4
 starknet-foundry 0.27.0
