+++ Scarb.toml
@@ -4,7 +4,7 @@ version = "0.1.0"
 edition = "2023_10"
 
 [dependencies]
-starknet = "2.8.0"
+starknet = "2.8.4"
 openzeppelin = { git = "https://github.com/OpenZeppelin/cairo-contracts.git", tag = "v0.16.0" }
 
 [dev-dependencies]
@@ -25,3 +25,5 @@ sierra = true
 [tool.fmt]
 sort-module-level-items = true
 
+[features]
+messaging_test = []
