+++ bindings/Cargo.toml
@@ -1,9 +1,16 @@
 [package]
-name = "bindings"
+name = "piltover"
 version = "0.1.0"
 edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
-cainome = { git = "https://github.com/cartridge-gg/cainome", rev = "v0.4.0", features = ["abigen-rs"] }
+cainome = { git = "https://github.com/cartridge-gg/cainome", rev = "v0.4.11", features = ["abigen-rs"] }
+starknet = "0.12"
+
+[lib]
+name = "piltover_bindings"
+path = "src/lib.rs"
+
+[[bin]]
+name = "generator"
+path = "src/bin.rs"
